refactor: phone-number-formatter 유틸화
