+++ pennyway-client-iOS/pennyway-client-iOS/View/LoginView/LoginView.swift
@@ -19,7 +19,7 @@ struct LoginView: View {
                             }
                         }
                 } else {
-                    LoginFormView(viewModel: LoginFormViewModel())
+                    LoginFormView(viewModel: LoginFormViewModel())   
                 }
             }
         }
